avatar displaying in world but not moving
