+++ client/3d/index.js
@@ -2,7 +2,7 @@ import {Vector3} from 'three'
 export function toKey(position) {
   return `${position.x},${position.y},${position.z}`
 }
-const removePeriod = coord => {
+/* const removePeriod = coord => {
   // more efficient way to remove a period? //
   return coord
     .toString()
@@ -13,7 +13,7 @@ export function cameraPositionToKey(position) {
   return `${removePeriod(position.x)},${removePeriod(
     position.y
   )},${removePeriod(position.z)}`
-}
+} */
 export function toPosition(key) {
   const positions = key.split(',')
   return new Vector3(positions[0], positions[1], positions[2])
