+++ client/3d/meshes/index.js
@@ -1,1 +1,2 @@
 export {makeUnitCube} from './makeUnitCube'
+export {makeHead, makeBody, makeArms, makeLegs} from './makeAvatar'
